requirements.txt
Crawl4AI
python-dotenv
certifi
pyaml
google-generativeai
openai
supabase
pydantic-ai
logfire
langgraph
langsmith
langchain_anthropic
langchain-google-genai
streamlit
PyPDF2
python-docx
llama-index
lama-index-llms-gemini
llama-index-embeddings-gemini
llama-parse

faiss-cpu